Add clim ids

+++ src/main/java/com/rbkmoney/newway/poller/event_stock/impl/party_mngmnt/party/PartyBlockingHandler.java
@@ -50,7 +50,7 @@ public class PartyBlockingHandler extends AbstractPartyManagementHandler {
         partySource.setRevision(null);
         partySource.setWtime(null);
         partySource.setEventId(eventId);
-        partySource.setSequenceId((long) event.getSequence());
+        partySource.setSequenceId(event.getSequence());
         partySource.setChangeId(changeId);
         partySource.setEventCreatedAt(TypeUtil.stringToLocalDateTime(event.getCreatedAt()));
         partySource.setBlocking(TBaseUtil.unionFieldToEnum(partyBlocking, com.rbkmoney.newway.domain.enums.Blocking.class));
